Restructured file... fix compiling errors, adjusted the cmakelists.txt to reflect changes
