+++ include/Entities/Player.h
@@ -1,0 +1,16 @@
+#ifndef PLAYER_H
+#define PLAYER_H
+
+#include <SDL2/SDL.h>
+#include "Sprite.h"
+
+class Player : public Sprite {
+public:
+    // Class Basics
+    Player();
+
+    // Utility 
+    void enforceBounds();
+};
+
+#endif
