+++ include/Managers/GameManager.h
