--- include/Player.h
@@ -1,16 +1,0 @@
-#ifndef PLAYER_H
-#define PLAYER_H
-
-#include <SDL2/SDL.h>
-#include "Sprite.h"
-
-class Player : public Sprite {
-public:
-    // Class Basics
-    Player();
-
-    // Utility 
-    void enforceBounds();
-};
-
-#endif
